Initial commit
Already added a modified version of fastnoiselib to be compatible with burst as well as made some burst jobs.
Made a mesh generation tool to create the terrain height map. Made some data structures for biomes so they can be easily added in to the project without any programming knowledge. 
Made a particle material to show the colors of a mesh instead of having to add specific textures. This is a temporary solution as I'm not an artist.